split up all other components reducer test
