+++ src/features/latest-news/redux/__test__/reducer.test.js
@@ -1,0 +1,20 @@
+import { UPDATE_LATEST_NEWS } from "../constants";
+import { latestNewsReducer } from "../reducer";
+
+const createInitialState = { selectedLatestNews: null };
+describe("with an UPDATE_LATEST_NEWS action", () => {
+  let newState;
+
+  beforeAll(() => {
+    const initialState = {
+      ...createInitialState,
+      selectedLatestNews: ""
+    };
+    const action = { type: UPDATE_LATEST_NEWS, payload: "NEWS" };
+    newState = latestNewsReducer(initialState, action);
+  });
+
+  it('it updates the latest news to "NEWS"', () => {
+    expect(newState.selectedLatestNews).toBe("NEWS");
+  });
+});
